# feat(matcher): cache labels matcher regexp results