In some scenarios Alertmanager might use labels regexp matching heavily:
- Inhibition
- Silences
- Dispatch

Same alert label strings are matched against same regexp expressions repeatedly.

We have observed ~25% cumulative cpu time usage by `regexp.(*Regexp).MatchString` in captured CPU profiles.

This change implements a cache in labels matcher to store regexp results.

It increases the performance of silence queries ~39-47% in existing benchmarks. Similar performance improvements are expected in inhibit and dispatch when regexp matching is used.

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/alertmanager/silence
cpu: Apple M3 Pro
                                   │ silence-main.txt │         silence-cached.txt          │
                                   │      sec/op      │   sec/op     vs base                │
Mutes/1_silence_mutes_alert-12            540.7n ± 1%   540.5n ± 0%        ~ (p=0.271 n=10)
Mutes/10_silences_mute_alert-12           1.391µ ± 0%   1.402µ ± 1%   +0.83% (p=0.000 n=10)
Mutes/100_silences_mute_alert-12          8.626µ ± 2%   8.739µ ± 1%   +1.30% (p=0.001 n=10)
Mutes/1000_silences_mute_alert-12         102.1µ ± 1%   106.4µ ± 1%   +4.21% (p=0.000 n=10)
Mutes/10000_silences_mute_alert-12        1.121m ± 1%   1.117m ± 0%   -0.36% (p=0.004 n=10)
Query/100_silences-12                    14.413µ ± 0%   8.206µ ± 2%  -43.06% (p=0.000 n=10)
Query/1000_silences-12                   153.99µ ± 0%   93.09µ ± 1%  -39.54% (p=0.000 n=10)
Query/10000_silences-12                   1.979m ± 0%   1.060m ± 1%  -46.45% (p=0.000 n=10)
geomean                                   36.66µ        29.89µ       -18.46%

                                   │ silence-main.txt │          silence-cached.txt           │
                                   │       B/op       │     B/op      vs base                 │
Mutes/1_silence_mutes_alert-12           1.180Ki ± 0%   1.180Ki ± 0%       ~ (p=1.000 n=10) ¹
Mutes/10_silences_mute_alert-12          4.227Ki ± 0%   4.227Ki ± 0%       ~ (p=1.000 n=10) ¹
Mutes/100_silences_mute_alert-12         33.32Ki ± 0%   33.32Ki ± 0%       ~ (p=1.000 n=10) ¹
Mutes/1000_silences_mute_alert-12        304.7Ki ± 0%   304.7Ki ± 0%  +0.00% (p=0.000 n=10)
Mutes/10000_silences_mute_alert-12       3.526Mi ± 0%   3.526Mi ± 0%  +0.02% (p=0.000 n=10)
Query/100_silences-12                    4.753Ki ± 0%   4.750Ki ± 0%  -0.06% (p=0.000 n=10)
Query/1000_silences-12                   39.92Ki ± 0%   39.91Ki ± 0%  -0.04% (p=0.000 n=10)
Query/10000_silences-12                  523.8Ki ± 0%   523.7Ki ± 0%  -0.01% (p=0.000 n=10)
geomean                                  45.44Ki        45.43Ki       -0.01%
¹ all samples are equal

                                   │ silence-main.txt │          silence-cached.txt          │
                                   │    allocs/op     │  allocs/op   vs base                 │
Mutes/1_silence_mutes_alert-12             19.00 ± 0%    19.00 ± 0%       ~ (p=1.000 n=10) ¹
Mutes/10_silences_mute_alert-12            40.00 ± 0%    40.00 ± 0%       ~ (p=1.000 n=10) ¹
Mutes/100_silences_mute_alert-12           139.0 ± 0%    139.0 ± 0%       ~ (p=1.000 n=10) ¹
Mutes/1000_silences_mute_alert-12         1.050k ± 0%   1.050k ± 0%       ~ (p=1.000 n=10) ¹
Mutes/10000_silences_mute_alert-12        10.09k ± 0%   10.10k ± 0%  +0.09% (p=0.000 n=10)
Query/100_silences-12                      32.00 ± 0%    32.00 ± 0%       ~ (p=1.000 n=10) ¹
Query/1000_silences-12                     128.0 ± 0%    128.0 ± 0%       ~ (p=1.000 n=10) ¹
Query/10000_silences-12                   1.038k ± 0%   1.038k ± 0%       ~ (p=1.000 n=10) ¹
geomean                                    216.1         216.1       +0.01%
¹ all samples are equal
```

Signed-off-by: Siavash Safi <siavash@cloudflare.com>