@grobinson-grafana 2025-04-09
You never evict stale items from the cache, but matchers are long lived (consider the case of routes). The Alertmanager will just accumulate memory by growing `sync.Map` until it OOMs?