@grobinson-grafana 2025-04-09
> > You never evict stale items from the cache, but matchers are long lived (consider the case of routes). The Alertmanager will just accumulate memory by growing `sync.Map` until it OOMs?
> 
> This was my thinking as well, but then concluded that memory usage will map statically to the amount of items configured in am and prom config? The more items you have the more memory needed, or am i missing something 👀
> 
> Nevertheless tho, @siavashs i think adding metrics on cache hit, miss and entries would help mapping memory increase with cache items. So that if OOM's happen, we can pin it to this cache.

Memory usage will increase as the number of unique label sets over time increases. For example, if you have 100 label matchers in memory, but you only see 100 unique alerts over a 7 day period, then in the worst case you will cache 100 * 100 entries (1000 entries).

If, on the other hand, you have the same number of label matchers in memory, but churn 1 million of unique alerts over a 7 day period (consider Grafana Mimir), then you will have 100 * 1 million = 100 million entries cached in memory, and you will never evict them.