@grobinson-grafana 2025-04-09
> > > > You never evict stale items from the cache, but matchers are long lived (consider the case of routes). The Alertmanager will just accumulate memory by growing `sync.Map` until it OOMs?
> > > 
> > > 
> > > This was my thinking as well, but then concluded that memory usage will map statically to the amount of items configured in am and prom config? The more items you have the more memory needed, or am i missing something 👀
> > > Nevertheless tho, @siavashs i think adding metrics on cache hit, miss and entries would help mapping memory increase with cache items. So that if OOM's happen, we can pin it to this cache.
> > 
> > 
> > Memory usage will increase as the number of unique label sets over time increases. For example, if you have 100 label matchers in memory, but you only see 100 unique alerts over a 24 hour period, then in the worst case you will cache 100 * 100 entries (10000 entries) in those 24 hours.
> > If, on the other hand, you have the same number of label matchers in memory, but churn 1 million of unique alerts over a 24 hour period (consider large multi-tenant systems like Grafana Mimir), then you will cache 100 * 1 million = 100 million entries in memory in those 24 hours. As you never evict them, it will keep accumulating until you OOM.
> 
> What is your suggestion for cache expiry? should we use a hard (configurable) limit like 1000 entries or use last hit time and/or number of hits to evict stale items?

Well, I think some analysis is required here.

With a hard limit like 1000 entries and no expiration, over time your cache could turn into a 100% miss rate as the data changes over time.

With something like LRU, I suspect you will find very high churn when you have lots of alerts as Prometheus will resend all of its alerts to the Alertmanager at regular intervals. I anticipate this will manifest as high cache miss rate, but feel free to run some experiments here.

Probably an LFU cache makes the most sense as you want to cache the most frequent inputs.

You may also find different access patterns for dispatching, silences and inhibitions, as these are run at different times.

What's the problem you are trying to solve our of interest? It's it something specific like silences, or just general regex matching?